+++ data_integration/config.py
@@ -72,25 +72,14 @@ def slack_token() -> typing.Optional[str]:
     """
     return None
 
-def teams_token() -> str:
-    """
-    When not None, then this teams webhook is notified of failed nodes.
-    """
-    return None
-
 @functools.lru_cache(maxsize=None)
 def event_handlers() -> [events.EventHandler]:
-    from .notification import notifier
-    chat_rooms = []
+    """User specific event handlers (mainly to notify chat systems)"""
+    configured_handlers = []
     if slack_token():
-        from .notification import slack
-        chat_rooms.append(slack.Slack())
-
-    if teams_token():
-        from .notification import teams
-        chat_rooms.append(teams.Teams())
-
-    return [notifier.Notifier(chat_rooms=chat_rooms)]
+        from data_integration.notification import slack
+        configured_handlers.append(slack.Slack())
+    return configured_handlers
 
 def password_masks() -> typing.List[str]:
     """Any passwords which should be masked in the UI or logs"""
